fix: upgrade to ConsoleAppFramework 5.7.13 and .NET 8
- Upgrade BLIT.CLI target framework from net6.0 to net8.0
- Set LangVersion to 13 to support ConsoleAppFramework 5.7.13 requirements
- Update command classes to use new API:
  - Remove inheritance from ConsoleAppBase
  - Rename classes to follow command naming convention
  - Replace [Option] attributes with document comments
  - Add XML documentation for better help text generation
- Update Program.cs to use new ConsoleApp.Create() API
- All modules now properly recognize and execute commands with new framework

+++ BLIT.Win/Pages/BannerIcons/Models/BannerIconEntry.cs
@@ -49,7 +49,7 @@ public class BannerIconEntry : BindableBase {
             OnPropertyChanged(nameof(AtlasName));
         }
     }
-    public int AtlasIndex => CellIndex / (TextureMerger.ROWS * TextureMerger.COLS);
+    public int AtlasIndex => CellIndex / (int)(TextureMerger.ROWS * TextureMerger.COLS);
 
     public string AtlasName => BannerUtils.GetAtlasName(_groupViewModel.GroupID, AtlasIndex);
     public int ID => BannerUtils.GetIconID(_groupViewModel.GroupID, CellIndex);
